# apps/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: TraceChain API
  version: 1.0.0
  description: |
    Off-chain API for the TraceChain pharmaceutical hybrid DApp. The API manages
    wallet-signature authentication, organization profiles, batch metadata,
    telemetry ingestion, document uploads, and blockchain event indexing.
servers:
  - url: http://localhost:4000
paths:
...
  /api/batches/draft:
    post:
      summary: Create an off-chain batch draft and return deterministic hashes for on-chain registration
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Draft created
  /api/batches/{id}/link-onchain:
    post:
      summary: Attach an on-chain batch ID and registration transaction hash to a draft
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Linked batch
  /api/batches/{identifier}:
    get:
      summary: Fetch a batch detail view by DB id, batch code, or on-chain batch ID
      parameters:
        - in: path
          name: identifier
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Batch detail
...
  /api/batches/{id}/transfers:
    post:
      summary: Log an off-chain transfer record after submitting the on-chain transaction
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Transfer logged
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT